// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-part-line-country-origin-model.ts
export class RFQSupplierPartLineCountryOriginModel {
    id: string;
    rfqSupplierHeaderInformationId: string;
    countryId: string;
    cityId: string;
    portId: string;
    ifsIncoTermId: string;
    incoTermCountryId: string;
    incoTermCityId: string;
    incoTermPortId: string;
    rfqSupplierPartLineId: string;
    countryName: string;
    cityName: string;
    portName: string;
    ifsIncoTermName: string;
}