// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-part-line-note-model.ts
export class RFQSupplierPartLineNoteModel {
    id: string;
    description: string;
    buyDescription: string;
    rfqId: string;
    supplierId: string;
    rfqPartLineNoteId: string;
    surrogateById: string;
    surrogationDate: string;
    attributeCategoryId: string;
    comment: string;
    supplierResponse: string;
    rfqSupplierPartLineId: string;
}