// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-partline.ts
import { RFXSupplierPartLineSubstitueModel } from './rfx-supplier-partline-substitue-model';

export class RFQSupplierPartLine {
    id: string;
    prNumber: string;
    lineNumber: string;
    partId: string;
    partDescirption: string;
    purchaseQty: any;
    purchaseUoMId: string;
    purchaseUoMCode: string;
    finalUnitPrice: any;
    totalPrice: any;
    discountPrice: string;
    unitPrice: any;
    discount: any;
    totalAmount: string;
    manufacter: string;
    originId: string;
    wantedDate: string;
    promisedDate: string;
    leadTime: string;
    deviation: string;
    showPriceToSuppliers: boolean;
    rfqId: string;
    supplierID: string;
    rfqParentId: string;
    isOverride: boolean;
    surrogateById: string;
    surrogateBy: string;
    surrogationDate: string;
    RfqPartLineId: string;
    attributeCategoryId: string;
    currencyId: string;
    currencyName: string;
    discountedValue: any;
    attachmentId: string;
    isDeleteandDownloadEnabledVisible: boolean;
    fileName: string;
    comment: string;
    rfqPartLineModel: any;
    listOfCurrency: any[];
    listOfCountries: any[];
    listOfIFSUoM: any[];
    listOfParts: any[];
    rfqSupplierPartLineAttributes: any[];
    rfqSupplierPartLineAttachments: any[];
    rfqSupplierPartLineCostFactors: any[];
    rfqSupplierPartLineDeliverables: any[];
    rfqSupplierPartLineDocumentTexts: any[];
    rfqSupplierPartLinePaymentSchedule: any[];
    rfqSupplierPartLineNotes: any[];
    rfqSupplierPartLineCountryOrigins: any[];
    rfxSupplierPartLineSubstitueModels: RFXSupplierPartLineSubstitueModel[];
    rfqSupplierPartLineCountryOriginModel: any[];
}