// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-partline-document-text-model.ts
export class RFQSupplierPartLineDocumentTextViewModel {
    id: string;
    rfqPartLineDocumentTextId: string;
    rfqPartLineDocumentTextModel: any;
    rFQId: string; 
    supplierID: string; 
    comments: string; 
    surrogateById: string; 
    surrogationDate: string; 
    supplierResponse: string; 
    rfqSupplierPartLineId: string;
}